Update structure of the directories to organize better
