+++ protocol-specs/generators/bft_processing/11_delegates_partial_switch.csv
