+++ protocol-specs/generators/bft_processing/4_delegates_simple.csv
