+++ protocol-specs/generators/bft_processing/7_delegates_partial_switch.csv
