+++ protocol-specs/generators/bft_processing/index.js
