+++ protocol-specs/generators/block_processing_transfers/index.js
@@ -15,7 +15,7 @@
 'use strict';
 
 const BaseGenerator = require('../base_generator');
-const defaultConfig = require('../../../config/devnet');
+const defaultConfig = require('../../config/devnet');
 const ChainStateBuilder = require('../../utils/chain_state_builder');
 
 const { genesisBlock } = defaultConfig;
@@ -81,7 +81,7 @@ const accounts = {
 			'iterations=1&salt=e8c7dae4c893e458e0ebb8bff9a36d84&cipherText=c0fab123d83c386ffacef9a171b6e0e0e9d913e58b7972df8e5ef358afbc65f99c9a2b6fe7716f708166ed72f59f007d2f96a91f48f0428dd51d7c9962e0c6a5fc27ca0722038f1f2cf16333&iv=1a2206e426c714091b7e48f6&tag=3a9d9f9f9a92c9a58296b8df64820c15&version=1',
 		password: 'elephant tree paris dragon chair galaxy',
 	},
-	existingDelegate: {
+	account: {
 		address: '10881167371402274308L',
 		publicKey:
 			'addb0e15a44b0fdc6ff291be28d8c98f5551d0cd9218d749e30ddb87c6e31ca9',
@@ -90,7 +90,7 @@ const accounts = {
 		balance: '0',
 		delegateName: 'genesis_100',
 	},
-	futureDelegate: {
+	secondAccount: {
 		passphrase:
 			'blame address tube insect cost knock major level regret bring april stick',
 		privateKey:
@@ -100,9 +100,18 @@ const accounts = {
 		address: '2222471382442610527L',
 		balance: '0',
 	},
+	thirdAccount: {
+		passphrase:
+			'female favorite client offer winner loud ostrich rich slogan jacket owner february',
+		privateKey:
+			'9b10f230fb2109cc77ee1a1ad659bf51f0e7b3c16697da416004dd409b427578ae88a228d37de7438f3d74df1713f3da00dd1c51afdc97bd47fa3aceb111a47a',
+		publicKey:
+			'ae88a228d37de7438f3d74df1713f3da00dd1c51afdc97bd47fa3aceb111a47a',
+		address: '11325618463998518034L',
+	},
 };
 
-const generateTestCasesValidBlockDelegateRegistration = () => {
+const generateTestCasesValidBlockTransferTx = () => {
 	const chainStateBuilder = new ChainStateBuilder(
 		genesisBlock,
 		initialAccountsState,
@@ -116,14 +125,9 @@ const generateTestCasesValidBlockDelegateRegistration = () => {
 		.forge();
 
 	chainStateBuilder
-		.transfer('30')
+		.transfer('40')
 		.from('10881167371402274308L')
 		.to('2222471382442610527L')
-		.forge();
-
-	chainStateBuilder
-		.registerDelegate('RadioHead')
-		.for('2222471382442610527L')
 		.forge();
 
 	const chainAndAccountStates = chainStateBuilder.getScenario();
@@ -131,20 +135,22 @@ const generateTestCasesValidBlockDelegateRegistration = () => {
 	return {
 		config: {
 			initialState: {
-				chain: chainAndAccountStates.chain.slice(0, 2),
+				// Given the library chainStateBuilder saves all mutations we use slice here to pick the first accounts state
+				chain: chainAndAccountStates.chain.slice(0, 1),
 				accounts: chainAndAccountStates.initialAccountsState,
 			},
 		},
-		description: 'A valid block with a delegate registration',
-		input: chainAndAccountStates.chain.slice(2)[0],
+		description: 'A valid block with a transfer transaction is processed',
+		input: chainAndAccountStates.chain.slice(1)[0],
 		output: {
 			chain: chainAndAccountStates.chain,
+			// Given the library chainStateBuilder saves all mutations we use slice here to pick the last account state
 			accounts: chainAndAccountStates.finalAccountsState.slice(-1),
 		},
 	};
 };
 
-const generateTestCasesInvalidBlockDelegateRegistrationSecondTime = () => {
+const generateTestCasesValidTransfersInvalidInSame = () => {
 	const chainStateBuilder = new ChainStateBuilder(
 		genesisBlock,
 		initialAccountsState,
@@ -158,19 +164,12 @@ const generateTestCasesInvalidBlockDelegateRegistrationSecondTime = () => {
 		.forge();
 
 	chainStateBuilder
-		.transfer('30')
+		.transfer('40')
 		.from('10881167371402274308L')
 		.to('2222471382442610527L')
-		.forge();
-
-	chainStateBuilder
-		.registerDelegate('RadioHead')
-		.for('2222471382442610527L')
-		.forge();
-
-	chainStateBuilder
-		.registerDelegate('RadioHead')
-		.for('2222471382442610527L')
+		.transfer('20')
+		.from('2222471382442610527L')
+		.to('11325618463998518034L')
 		.forgeInvalidInputBlock();
 
 	const chainAndAccountStates = chainStateBuilder.getScenario();
@@ -178,54 +177,12 @@ const generateTestCasesInvalidBlockDelegateRegistrationSecondTime = () => {
 	return {
 		config: {
 			initialState: {
-				chain: chainAndAccountStates.chain.slice(0, 3),
-				accounts: chainAndAccountStates.initialAccountsState,
-			},
-		},
-		description: 'An invalid block with a second delegate registration',
-		input: chainAndAccountStates.inputBlock[0],
-		output: {
-			chain: chainAndAccountStates.chain,
-			accounts: chainAndAccountStates.finalAccountsState.slice(-1),
-		},
-	};
-};
-
-const generateTestCasesInvalidBlockDelegateRegistrationForbiddenName = () => {
-	const chainStateBuilder = new ChainStateBuilder(
-		genesisBlock,
-		initialAccountsState,
-		accounts,
-	);
-
-	chainStateBuilder
-		.transfer('50')
-		.from('16313739661670634666L')
-		.to('10881167371402274308L')
-		.forge();
-
-	chainStateBuilder
-		.transfer('30')
-		.from('10881167371402274308L')
-		.to('2222471382442610527L')
-		.forge();
-
-	chainStateBuilder
-		.registerDelegate('2222471382442610527L')
-		.for('2222471382442610527L')
-		.forgeInvalidInputBlock();
-
-	const chainAndAccountStates = chainStateBuilder.getScenario();
-
-	return {
-		config: {
-			initialState: {
-				chain: chainAndAccountStates.chain.slice(0, 2),
+				chain: chainAndAccountStates.chain.slice(0, 1),
 				accounts: chainAndAccountStates.initialAccountsState,
 			},
 		},
 		description:
-			'An invalid block with a delegate registration using invalid name',
+			'An invalid block with transfers valid on their own but invalid in the context of same block',
 		input: chainAndAccountStates.inputBlock[0],
 		output: {
 			chain: chainAndAccountStates.chain,
@@ -234,35 +191,85 @@ const generateTestCasesInvalidBlockDelegateRegistrationForbiddenName = () => {
 	};
 };
 
-const validBlockWithDelegateRegistrationSuite = () => ({
+const generateTestCasesTransferTooMuchSpentInBlockContext = () => {
+	const chainStateBuilder = new ChainStateBuilder(
+		genesisBlock,
+		initialAccountsState,
+		accounts,
+	);
+
+	chainStateBuilder
+		.transfer('50')
+		.from('16313739661670634666L')
+		.to('10881167371402274308L')
+		.forge();
+
+	chainStateBuilder
+		.transfer('1')
+		.from('10881167371402274308L')
+		.to('2222471382442610527L')
+		.forge();
+
+	// After the first transfer of 0.5 lsk (0.5 + 0.1 fee) the second transfer leaves the account with insufficient funds (0.4)
+	chainStateBuilder
+		.transfer('0.5')
+		.from('2222471382442610527L')
+		.to('10881167371402274308L')
+		.transfer('0.5')
+		.from('2222471382442610527L')
+		.to('11325618463998518034L')
+		.forgeInvalidInputBlock();
+
+	const chainAndAccountStates = chainStateBuilder.getScenario();
+
+	return {
+		config: {
+			initialState: {
+				chain: chainAndAccountStates.chain.slice(0, 1),
+				accounts: chainAndAccountStates.initialAccountsState,
+			},
+		},
+		description:
+			'An invalid block with transfers valid on their own but second transfer would not have enough funds after fee is applied',
+		input: chainAndAccountStates.inputBlock[0],
+		output: {
+			chain: chainAndAccountStates.chain,
+			accounts: chainAndAccountStates.finalAccountsState.slice(-1),
+		},
+	};
+};
+
+const validBlockWithTransferTxSuite = () => ({
 	title: 'Valid block processing',
-	summary: 'A valid block with a delegate registration',
+	summary: 'A valid block with a transfer transaction is processed',
 	config: { network: 'mainnet' },
-	runner: 'block_processing_delegate',
-	handler: 'valid_block_processing_delegate_registration_tx',
-	testCases: [generateTestCasesValidBlockDelegateRegistration()],
+	runner: 'block_processing_transfers',
+	handler: 'valid_block_processing_one_transfer_tx',
+	testCases: [generateTestCasesValidBlockTransferTx()],
 });
 
-const invalidBlockWithSecondDelegateRegistrationSuite = () => ({
+const invalidBlockFundingAndTransferSameBlock = () => ({
 	title: 'Invalid block processing',
-	summary: 'An invalid block with a second delegate registration',
+	summary:
+		'An invalid block with transfers valid on their own but invalid in the context of same block',
 	config: { network: 'mainnet' },
-	runner: 'block_processing_delegate',
-	handler: 'invalid_block_processing_second_delegate_registration_tx',
-	testCases: [generateTestCasesInvalidBlockDelegateRegistrationSecondTime()],
+	runner: 'block_processing_transfers',
+	handler: 'invalid_block_processing_funding_and_transfer_same_block',
+	testCases: [generateTestCasesValidTransfersInvalidInSame()],
 });
 
-const invalidBlockWithForbiddenNameDelegateRegistrationSuite = () => ({
+const invalidBlockTooMuchSpent = () => ({
 	title: 'Invalid block processing',
-	summary: 'An invalid block with a delegate registration using invalid name',
+	summary:
+		'An invalid block with transfers valid on their own but second transfer would not have enough funds after fee is applied',
 	config: { network: 'mainnet' },
-	runner: 'block_processing_delegate',
-	handler: 'invalid_block_processing_forbidden_name_delegate_registration_tx',
-	testCases: [generateTestCasesInvalidBlockDelegateRegistrationForbiddenName()],
+	runner: 'block_processing_transfers',
+	handler: 'invalid_block_processing_not_enough_balance_for_second_transaction',
+	testCases: [generateTestCasesTransferTooMuchSpentInBlockContext()],
 });
 
-module.exports = BaseGenerator.runGenerator('block_processing_delegate', [
-	validBlockWithDelegateRegistrationSuite,
-	invalidBlockWithSecondDelegateRegistrationSuite,
-	invalidBlockWithForbiddenNameDelegateRegistrationSuite,
+module.exports = BaseGenerator.runGenerator('block_processing_transfers', [
+	validBlockWithTransferTxSuite,
+	invalidBlockFundingAndTransferSameBlock,
+	invalidBlockTooMuchSpent,
 ]);
