+++ protocol-specs/generators/transaction_signing/index.js
