+++ protocol-specs/package.json
@@ -7,8 +7,8 @@
 		"lint": "eslint .",
 		"lint:fix": "eslint --fix .",
 		"format": "prettier --write '**/*'",
-		"generate-all-specs": "node src/index.js",
-		"validate-all-specs": "node src/validate.js"
+		"generate-all-specs": "node scripts/generate.js",
+		"validate-all-specs": "node scripts/validate.js"
 	},
 	"author": "Lisk Foundation <admin@lisk.io>, lightcurve GmbH <admin@lightcurve.io>",
 	"license": "Apache-2.0",
