--- protocol-specs/src/spec_schema/chain_state.js
@@ -1,38 +1,0 @@
-/*
- * Copyright © 2018 Lisk Foundation
- *
- * See the LICENSE file at the top-level directory of this distribution
- * for licensing information.
- *
- * Unless otherwise agreed in a custom licensing agreement with the Lisk Foundation,
- * no part of this software, including this file, may be copied, modified,
- * propagated, or distributed except according to the terms contained in the
- * LICENSE file.
- *
- * Removal or modification of this copyright notice is prohibited.
- */
-
-'use strict';
-
-const blockSchema = require('./block');
-const accountSchema = require('./account');
-
-module.exports = {
-	type: 'object',
-	properties: {
-		chain: {
-			type: 'array',
-			uniqueItems: true,
-			items: {
-				...blockSchema,
-			},
-		},
-		accounts: {
-			type: 'array',
-			uniqueItems: true,
-			items: {
-				...accountSchema,
-			},
-		},
-	},
-};
