+++ protocol-specs/utils/bft/index.js
