+++ protocol-specs/utils/blocks/block.js
