+++ protocol-specs/utils/dpos/index.js
